requirements.txt
qgate-perf==0.4.7
qgate-graph==1.4.6
cassandra-driver==3.29.1

# https://github.com/theskumar/python-dotenv
python-dotenv~=1.0
